backend/requirements.txt
fastapi
uvicorn
pandas
numpy
nltk
fasttext
scikit-learn
transformers==4.36.2
dialogtag
regex
python-multipart
tensorflow-cpu==2.15.0
llama-cpp-python
python-dotenv
requests
google-generativeai
openai>=1.0.0
gdown
requests
tqdm